API/requirements.txt
groq
python-dotenv
PyPDF2
langchain
langchain_groq
langchain_pinecone
python-dotenv
pymupdf
sentence-transformers
langchain-community
uvicorn
FastAPI
-e .